Moved/copied low-level tests to _libs/tests
