+++ ucp/_libs/tests/test_peer_send_recv.py
@@ -1,6 +1,5 @@
 import multiprocessing as mp
 import os
-import pickle
 import time
 
 import pytest
@@ -9,18 +8,6 @@ from ucp._libs import ucx_api
 from ucp._libs.arr import Array
 
 mp = mp.get_context("spawn")
-
-
-def test_pickle_ucx_address():
-    ctx = ucx_api.UCXContext()
-    worker = ucx_api.UCXWorker(ctx)
-    org_address = worker.get_address()
-    dumped_address = pickle.dumps(org_address)
-    org_address_hash = hash(org_address)
-    org_address = bytes(org_address)
-    new_address = pickle.loads(dumped_address)
-    assert org_address_hash == hash(new_address)
-    assert bytes(org_address) == bytes(new_address)
 
 
 def blocking_handler(request, exception, finished):
